ausa_edge_solution/ausa/modules/flutter_subscriber/requirements.txt
amqtt==0.10.1
requests==2.31.0
azure-iot-device~=2.7.0
cryptography==45.0.3
httpx[http2]==0.24.1